Added create and delete vpc functions. These are working.

+++ basic-ec2/scenario1.sh
@@ -6,35 +6,61 @@
 #
 # https://docs.aws.amazon.com/vpc/latest/userguide/VPC_Scenario1.html
 # 
-
 # This will be a script to create and destroy the resources in the documentation.
 # It will use an idempotent approach, only creating what doesn't exist.
-# For now, it's just a collection of useful commands.
-# 
+#
+# Ansible or Terraform would be a better tool. This is just for learning.
+
+
+create-vpc() {
+	echo Create
+	# Check if Vpc exists
+	VPCID=`aws ec2 describe-vpcs --output text --query 'Vpcs[?Tags[?Key==\`example\` && Value==\`sn1\`]].VpcId'`
+	if [ $? -gt 0 ] ; then
+		echo "describe-vpcs command error" >&2
+		exit 1
+	fi
+	if [ -z "$VPCID" ] ; then
+		echo "Creating VPC"
+		# if vpc does not exist, create it
+		#aws ec2 create-vpc --cridr-block 10.0.0.0/16
+		VPCID=`aws ec2 create-vpc --cidr-block 10.0.0.0/16 --output text --query 'Vpc.VpcId'`
+		aws ec2 create-tags --resources $VPCID --tags Key=example,Value=sn1		
+	fi
+	echo "VpcId=$VPCID."
+	#aws ec2 describe-vpcs
+}
+
+
+delete-vpc() {
+	echo Destroy
+	# Now tear down the VPC.
+	VPCID=`aws ec2 describe-vpcs --output text --query 'Vpcs[?Tags[?Key==\`example\` && Value==\`sn1\`]].VpcId'`
+	if [ $? -gt 0 ] ; then
+		echo "describe-vpcs command error" >&2
+		exit 1
+	fi
+
+	# VpcId found, now delete it
+	if [ -n "$VPCID" ] ; then
+		echo "Deleting VPC"
+		aws ec2 delete-vpc --vpc-id $VPCID
+	fi
+
+	# check if deleted
+	VPCID=`aws ec2 describe-vpcs --output text --query 'Vpcs[?Tags[?Key==\`example\` && Value==\`sn1\`]].VpcId'`
+	echo "VpcId=$VPCID."
+	#aws ec2 describe-vpcs
+}
+
 
 #
-# The create sequence
+# Start point
 #
-
-# Check if Vpc exists
-VPCID=`aws ec2 describe-vpcs --output text --query 'Vpcs[?Tags[?Key==\`example\` && Value==\`sn1\`]].VpcId'
-# todo - check the result
-
-# if vpc does not exist, create it
-aws ec2 create-vpc --cridr-block 10.0.0.0/16
-aws ec2 create-tags --resources $VPCID --tags Key=example,Value=sn1
-VPCID=`aws ec2 create-vpc --cidr-block 10.0.0.0/16 --output text --query 'Vpc.VpcId'`
-
-
-echo $VPCID
-#aws ec2 describe-vpcs
-
-
-#
-# The destroy sequence
-#
-
-# Now tear down the VPC.
-aws ec2 delete-vpc --vpc-id $VPCID
-
-
+case "$1" in
+	cv) create-vpc ;;
+	dv) delete-vpc ;;
+	*)
+		echo "Usage: $0 {create-vpc|delete-vpc}"
+		exit 1
+esac
